refactor: shared webview dom helpers and html scaffold
